Make CCryptoKeyStore::Unlock check all keys.
CCryptoKeyStore::Unlock has a loop to attempt decrypting each key which only executes once, likely due to a simple mistake when the code was originally written.

This patch fixes the behavior by making it check all keys. It also adds
a fatal assertion in the case some decrypt but some do not, since that indicates that the wallet is in some kind of really bad state.

This may make unlocking noticeably slower on wallets with many keys.

https://github.com/bitcoin/bitcoin/commit/1e21c17d208e310295475c0e4a46d750a5c9ba2d

Dont run full check every time we decrypt wallet.

https://github.com/bitcoin/bitcoin/commit/a35b55b522da8eeaed895f0ed43ba595fc083498