Bugfix: remove conflicting transactions from memory pool.
When a transaction A is in the memory pool, while a transaction B
(which shares an input with A) gets accepted into a block, A was
kept forever in the memory pool.

This commit adds a CTxMemPool::removeConflicts method, which
removes transactions that conflict with a given transaction, and
all their children.

This results in less transactions in the memory pool, and faster
construction of new blocks.

https://github.com/bitcoin/bitcoin/commit/231b399952fd620ee0f72b1947024dba9651630d