Reimplement CBlockLocator's chain-related logic in CChain,
This removes a few unused CBlockLocator methods, and moves the
construction and fork-finding logic to CChain (which can do these
more efficiently, as it has a height-indexable chain available).
It also makes CBlockLocator independent from the validation code.
https://github.com/bitcoin/bitcoin/commit/e4daecda0bcd47a2672eb625232f00e388a3cd87

init.cpp: Bugfix: if no bestblock record is present, do a -rescan.
It is possible to have a wallet.dat file without any bestblock
record at all (if created offline, for example), which - when
loaded into a client with a up-to-date chain - does no rescan and
shows no transactions.
Also make sure to write the current best block after a rescan, so
it isn't necessary twice.
https://github.com/bitcoin/bitcoin/commit/2aceeb01a9b0b0c6bfa0d56211453e87bc33d6e4